feat(CodeArts): move codearts inspector and deploy to new directory
