+++ huaweicloud/services/acceptance/codeartsdeploy/resource_huaweicloud_codearts_deploy_host_test.go
@@ -1,4 +1,4 @@
-package codearts
+package codeartsdeploy
 
 import (
 	"fmt"
