+++ huaweicloud/services/acceptance/codeartsinspector/resource_huaweicloud_codearts_inspector_website_scan_test.go
@@ -1,4 +1,4 @@
-package codearts
+package codeartsinspector
 
 import (
 	"fmt"
