+++ huaweicloud/services/codearts/resource_huaweicloud_codearts_project.go
@@ -1,8 +1,3 @@
-// ---------------------------------------------------------------
-// *** AUTO GENERATED CODE ***
-// @Product ProjectMan
-// ---------------------------------------------------------------
-
 package codearts
 
 import (
