+++ huaweicloud/services/codearts/resource_huaweicloud_codearts_repository.go
@@ -1,8 +1,3 @@
-// ---------------------------------------------------------------
-// *** AUTO GENERATED CODE ***
-// @Product CodeHub
-// ---------------------------------------------------------------
-
 package codearts
 
 import (
